pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "poker_teaching"
version = "0.4.0"
description = "Poker Teaching System - Django First (v0.4)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0,<6",
    "djangorestframework>=3.15",
    "drf-spectacular>=0.27",
    "psycopg[binary]>=3.1",
    "prometheus_client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "coverage>=7.5",
    "pytest-django>=4.8",
    "openapi-spec-validator>=0.7",
    "ruff>=0.5",
    "black>=24.4",
    "isort>=5.13",
    "mypy>=1.10",
    "factory-boy>=3.3",
    "pre-commit>=3.7",
    "types-requests",
    "mkdocs>=1.6",
    "pokerkit==0.6.3",
]

[tool.setuptools]
package-dir = {"" = "packages"}

[tool.setuptools.packages.find]
where = ["packages"]
include = ["poker_core*"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["E","F","I","UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.coverage.report]
show_missing = true
